large refactor for modularity so "@/eos" and "@/novusphere" can be used in other projects
