+++ src/novusphere/index.js
@@ -1,10 +1,9 @@
-import { NovusphereAdapter } from "@/novusphere-adapter"
-import { storage, SaveStorage } from "@/storage"
+import { NovusphereAdapter } from "./adapter";
+import { storage, SaveStorage } from "@/storage";
 
 const NovusphereConfig = {
-    url: 'https://db.novusphere.io', // point to your own instance of novusphere-db
-    //url: 'http://localhost:8099',
-    collection: 'eosforum',
+    url: 'https://db.novusphere.io',
+    collection_forum: 'eosforum',
     collection_post_state: 'ns_post_state',
     collection_post_vote: 'ns_post_vote'
 }
