+++ src/novusphere/query/skip.js
@@ -1,0 +1,5 @@
+export default {
+    page(current, maxItems) {
+        return (current - 1) * maxItems;
+    }
+}
