+++ src/requests.js
@@ -1,15 +1,7 @@
 import jQuery from "jquery";
 
 export default {
-    GetHost(href) {
-        if (href.indexOf("magnet:") == 0) {
-            return "magnet link";
-        }
-        var parser = document.createElement("a");
-        parser.href = href;
-        return parser.host.toLowerCase();
-    },
-    AsyncPost(url, data) {
+    post(url, data) {
         return new Promise((resolve, reject) => {
             jQuery.post(url, data, function(result) {
                 resolve(result);
@@ -19,7 +11,7 @@ export default {
             });
         });
     },
-    AsyncGet(url, data) {
+    get(url, data) {
         return new Promise((resolve, reject) => {
             jQuery.get(url, data, function(result) {
                 resolve(result);
